# app/src/api/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'http://localhost:4000' | (string & {});
};

export type Note = {
    /**
     * Unique identifier
     */
    id: string;
    /**
     * Note title
     */
    title: string;
    /**
     * Note body
     */
    content: string;
    /**
     * Whether the note is completed
     */
    completed: boolean;
    /**
     * Creation timestamp
     */
    createdAt: string;
    /**
     * Last update timestamp
     */
    updatedAt: string;
};

export type CreateNote = {
    /**
     * Note title
     */
    title: string;
    /**
     * Note body
     */
    content: string;
    /**
     * Whether the note is completed
     */
    completed?: boolean;
};

export type UpdateNote = {
    /**
     * Note title
     */
    title?: string;
    /**
     * Note body
     */
    content?: string;
    /**
     * Whether the note is completed
     */
    completed?: boolean;
};

export type NoteList = {
    items: Array<{
        /**
         * Unique identifier
         */
        id: string;
        /**
         * Note title
         */
        title: string;
        /**
         * Note body
         */
        content: string;
        /**
         * Whether the note is completed
         */
        completed: boolean;
        /**
         * Creation timestamp
         */
        createdAt: string;
        /**
         * Last update timestamp
         */
        updatedAt: string;
    }>;
    /**
     * Total number of notes
     */
    total: number;
};

export type Error = {
    /**
     * Error message
     */
    error: string;
};

/**
 * Unique identifier
 */
export type NoteId = string;

/**
 * Unique identifier
 */
export type NoteId2 = NoteId;

export type GetApiNotesData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/notes';
};

export type GetApiNotesResponses = {
    /**
     * List of notes
     */
    200: {
        items: Array<{
            /**
             * Unique identifier
             */
            id: string;
            /**
             * Note title
             */
            title: string;
            /**
             * Note body
             */
            content: string;
            /**
             * Whether the note is completed
             */
            completed: boolean;
            /**
             * Creation timestamp
             */
            createdAt: string;
            /**
             * Last update timestamp
             */
            updatedAt: string;
        }>;
        /**
         * Total number of notes
         */
        total: number;
    };
};

export type GetApiNotesResponse = GetApiNotesResponses[keyof GetApiNotesResponses];

export type PostApiNotesData = {
    body?: {
        /**
         * Note title
         */
        title: string;
        /**
         * Note body
         */
        content: string;
        /**
         * Whether the note is completed
         */
        completed?: boolean;
    };
    path?: never;
    query?: never;
    url: '/api/notes';
};

export type PostApiNotesErrors = {
    /**
     * Validation error
     */
    400: {
        /**
         * Error message
         */
        error: string;
    };
};

export type PostApiNotesError = PostApiNotesErrors[keyof PostApiNotesErrors];

export type PostApiNotesResponses = {
    /**
     * Note created
     */
    201: {
        /**
         * Unique identifier
         */
        id: string;
        /**
         * Note title
         */
        title: string;
        /**
         * Note body
         */
        content: string;
        /**
         * Whether the note is completed
         */
        completed: boolean;
        /**
         * Creation timestamp
         */
        createdAt: string;
        /**
         * Last update timestamp
         */
        updatedAt: string;
    };
};

export type PostApiNotesResponse = PostApiNotesResponses[keyof PostApiNotesResponses];

export type DeleteApiNotesByIdData = {
    body?: never;
    path: {
        /**
         * Unique identifier
         */
        id: string;
    };
    query?: never;
    url: '/api/notes/{id}';
};

export type DeleteApiNotesByIdErrors = {
    /**
     * Note not found
     */
    404: {
        /**
         * Error message
         */
        error: string;
    };
};

export type DeleteApiNotesByIdError = DeleteApiNotesByIdErrors[keyof DeleteApiNotesByIdErrors];

export type DeleteApiNotesByIdResponses = {
    /**
     * Note deleted
     */
    204: void;
};

export type DeleteApiNotesByIdResponse = DeleteApiNotesByIdResponses[keyof DeleteApiNotesByIdResponses];

export type GetApiNotesByIdData = {
    body?: never;
    path: {
        /**
         * Unique identifier
         */
        id: string;
    };
    query?: never;
    url: '/api/notes/{id}';
};

export type GetApiNotesByIdErrors = {
    /**
     * Note not found
     */
    404: {
        /**
         * Error message
         */
        error: string;
    };
};

export type GetApiNotesByIdError = GetApiNotesByIdErrors[keyof GetApiNotesByIdErrors];

export type GetApiNotesByIdResponses = {
    /**
     * Note found
     */
    200: {
        /**
         * Unique identifier
         */
        id: string;
        /**
         * Note title
         */
        title: string;
        /**
         * Note body
         */
        content: string;
        /**
         * Whether the note is completed
         */
        completed: boolean;
        /**
         * Creation timestamp
         */
        createdAt: string;
        /**
         * Last update timestamp
         */
        updatedAt: string;
    };
};

export type GetApiNotesByIdResponse = GetApiNotesByIdResponses[keyof GetApiNotesByIdResponses];

export type PutApiNotesByIdData = {
    body?: {
        /**
         * Note title
         */
        title?: string;
        /**
         * Note body
         */
        content?: string;
        /**
         * Whether the note is completed
         */
        completed?: boolean;
    };
    path: {
        /**
         * Unique identifier
         */
        id: string;
    };
    query?: never;
    url: '/api/notes/{id}';
};

export type PutApiNotesByIdErrors = {
    /**
     * Validation error
     */
    400: {
        /**
         * Error message
         */
        error: string;
    };
    /**
     * Note not found
     */
    404: {
        /**
         * Error message
         */
        error: string;
    };
};

export type PutApiNotesByIdError = PutApiNotesByIdErrors[keyof PutApiNotesByIdErrors];

export type PutApiNotesByIdResponses = {
    /**
     * Note updated
     */
    200: {
        /**
         * Unique identifier
         */
        id: string;
        /**
         * Note title
         */
        title: string;
        /**
         * Note body
         */
        content: string;
        /**
         * Whether the note is completed
         */
        completed: boolean;
        /**
         * Creation timestamp
         */
        createdAt: string;
        /**
         * Last update timestamp
         */
        updatedAt: string;
    };
};

export type PutApiNotesByIdResponse = PutApiNotesByIdResponses[keyof PutApiNotesByIdResponses];